Remove clean wc check in git feature-end
This check conflicts with git feature-push no longer changing the tracking branch by default.  Because of this, by default the local branch will continue to track upstream/develop by default.  Therefore, the local branch will, by default, show unpushed changes against it's tracking branch.

One option would be to change this check to use the derived branch name git feature-push uses (same branch name chopping of ticket number and versions).  However this would be fragile, because if the branch was pushed to a remote branch of a different name, this issue still would not be resolved.

I like this check, but until a better solution can be found I am removing this.  Anyone using git feature-end will just need to be care and know what they are doing!